% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pixelToPolygon.R
\name{pixelToPolygon}
\alias{pixelToPolygon}
\title{Convert pixel to polygon}
\usage{
pixelToPolygon(x, cell, plot = FALSE)
}
\arguments{
\item{x}{Raster* object}

\item{cell}{Cell number (numeric) or vector of length 2 representing the x and y coordinate of the desired pixel}
}
\value{
Object of class \code{SpatialPolygons} encompassing the outline of the pixel located at \code{cell}.
}
\description{
Convert a pixel to a polygon object by giving the raster object and cell number or xy coordinates vector
}

