% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stretchBrick.R
\name{stretchBrick}
\alias{stretchBrick}
\title{Multitemporal stretch}
\usage{
stretchBrick(x, inlo = min(minValue(x)), inup = max(maxValue(x)),
  outlo = 1, outup = 256, stretch = "lin", ...)
}
\arguments{
\item{x}{RasterBrick}
}
\description{
...work in progress...
}

