% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subsetScenes.R
\name{subsetScenes}
\alias{subsetScenes}
\title{Subset Scenes}
\usage{
subsetScenes(x, targ, padding = NULL, filename = "", verbose = TRUE, ...)
}
\arguments{
\item{x}{A spatial object (polygon, or otherwise) to base the subset extent on}

\item{targ}{A character vector of filenames, or a list of raster layers to be subsetted}

\item{filename}{Character. Optional: If results are to be written to file, then a character vector of \code{length = length(targ)} should be provided. Not supported yet (leave blank, or it will return an error)}

\item{verbose}{Logical. Print status messages to the console?}
}
\value{
A list of avialable rasters cropped to the extent of the input object + padding. Side effect: cropped rasters written to file if a filename vector is supplied (not supported yet).
}
\description{
Subset scenes from a list or directory based on a given spatial object. Takes results of \code{selectScenes()} and subsets based on intersecting extents.
}
\author{
Ben DeVries
}

