% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subsetScenesPlot.R
\name{subsetScenesPlot}
\alias{subsetScenesPlot}
\title{Plot subsetted scenes in sequence}
\usage{
subsetScenesPlot(..., bands = NULL, stretch = NULL, textcol = "white")
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{subsetScenes}}}

\item{bands}{Numeric. Bands to display if an RGB composite (ignored if \code{length(bands)} is less than 3)}

\item{stretch}{Character. Stretch to apply to composite (see \code{\link{plotRGB}} for more info)}

\item{textcol}{Character. Colour of text to display on image.}
}
\value{
list of raster subsets (see \code{\link{subsetScenes}})
}
\description{
Plot subsetted scenes in sequence. Same as \code{\link{subsetScenes}}, except that it plots the subsets afterwards.
}

