% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tsChips.R
\name{tsChips}
\alias{tsChips}
\title{Plot TS Chips}
\usage{
tsChips(x, loc, start = NULL, end = NULL, buff = 17, percNA = 20,
  cols = "PiYG", nbks = 35, nc = 3, nr = 3, ggplot = FALSE,
  export = FALSE, show = TRUE)
}
\arguments{
\item{x}{RasterBrick. Image time series brick}

\item{loc}{Location. Can be a vector of length 2 representing the x,y coordinates, or a SpatialPolygons or SpatialPoints object of \code{nrow = 1} (the first row will be taken if \code{nrow(loc) > 1}), or an extent object (which will be extended if a buffer > 0 is given; see below)}

\item{start}{Date. OptionaL: earliest date ("yyyy-dd-mm") to display.}

\item{end}{Date. Optional: latest date ("yyyy-dd-mm") to display.}

\item{buff}{Numeric. Number of pixels to buffer the location in all directions. A higher buffer will essentially zoom out.}

\item{percNA}{Numeric. Maximum allowable \% NA in the cropped image chips}

\item{cols}{Character. Name of colour map to use (see display.brewer.all()) or a character vector with two or more colour names or hexadecimal values (as strings) between which to interpolate.}

\item{nbks}{Numeric. Number of breaks in the colour map}

\item{ggplot}{Logical. Produce a ggplot time series plot object?}

\item{export}{Logical. Export processed chips to workspace as a rasterBrick? If \code{TRUE} and \code{ggplot = TRUE} as well, then both will be exported as a list object.}

\item{show}{Logical. Plot the chips? Can be set to \code{FALSE} if you just want to export the chips as rasterBrick with or without the ggplot object.}

\item{nc/nr}{Numeric. Number of columns and rows to plot, respectively. If the number of layers is greater than \code{nc*nr}, a screen prompt will lead to the next series of plots. These cannot exceed 4.}
}
\value{
\code{NULL} if \code{ggplot = FALSE} or an object of class \code{ggplot} if \code{ggplot = TRUE}, with the side effect of time series chips being plotted in both cases. If \code{export = TRUE}, an object of class rasterBrick, and if both \code{ggplot} and \code{export} are \code{TRUE}, a list including a rasterBrick and a ggplot object.
}
\description{
Plot image TS chips based on a location, area buffer, and time buffer
}
\examples{
\dontrun{
library(bfastSpatial)
data(tura)

tsChips(tura, loc = c(820796, 831198))
tsChips(tura, loc = c(820796, 831198), buff = 50) # zoom out
tsChips(tura, loc = c(820796, 831198), buff = 50, percNA = 80) # allow more NA's in field of view
tsChips(tura, loc = c(820796, 831198), start = "2007-01-01", end = "2012-01-01", ggplot = TRUE) # restrict dates and produce pixel time series plot afterwards

# store a ggplot object in workspace
p <- tsChips(tura, loc = c(820796, 831198), start = "1999-01-01", ggplot = TRUE)
# add layers to the plot
p <- p + geom_point(col = "blue")
p <- p + geom_line(col = "red", lty = 2)
p <- p + labs(y = "NDVI")
p <- p + scale_y_continuous(limits = c(0, 10000), breaks = c(0, 2500, 5000, 7500, 10000), labels = c("0.0", "0.25", "0.50", "0.75", "1.0"))

# alternative colour scales
library(RColorBrewer)
display.brewer.all()
tsChips(tura, loc = c(820796, 831198), start = "1999-01-01", cols = "Spectral")
tsChips(tura, loc = c(820796, 831198), start = "1999-01-01", cols = "Greens")

tsChips(tura, loc = c(820796, 831198), start = "1999-01-01", cols = c("red", "yellow", "blue"))
tsChips(tura, loc = c(820796, 831198), start = "1999-01-01", cols = c("#DEEBF7", "#3182BD"))

# draw a custom SpatialPoygons object and plot around that
plot(tura, 6)
pol <- drawPoly(sp = TRUE) # click 'Finish' in plot window when done
projection(pol) <- projection(tura)
plot(tura, 6); plot(pol, add=TRUE)
tsChips(tura, loc = pol, start = "1999-01-01", ggplot = TRUE)
}
}
\author{
Ben DeVries
}
\references{
Cohen, W. B., Yang, Z., Kennedy, R. (2010). Detecting trends in forest disturbance and recovery using yearly Landsat time series: 2. TimeSync - Tools for calibration and validation. Remote Sensing of Environment, 114(12), 2911-2924.
}

