% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tsChipsRGB.R
\name{tsChipsRGB}
\alias{tsChipsRGB}
\title{Plot TS Chips (RGB version)}
\usage{
tsChipsRGB(xr, xg, xb, loc, start = NULL, end = NULL, buff = 17,
  percNA = 20, nc = 3, nr = 3, ggplot = FALSE, cores = 1,
  textcol = "white", ...)
}
\arguments{
\item{loc}{Location. Can be a vector of length 2 representing the x,y coordinates, or a SpatialPolygons or SpatialPoints object of \code{nrow = 1} (the first row will be taken if \code{nrow(loc) > 1}), or an extent object (which will be extended if a buffer > 0 is given; see below)}

\item{start}{Date. OptionaL: earliest date ("yyyy-dd-mm") to display.}

\item{end}{Date. Optional: latest date ("yyyy-dd-mm") to display.}

\item{buff}{Numeric. Number of pixels to buffer the location in all directions. A higher buffer will essentially zoom out.}

\item{percNA}{Numeric. Maximum allowable \% NA in the cropped image chips}

\item{ggplot}{Logical. Produce a ggplot time series plot object?}

\item{cores}{Numeric. Number of cores to use for pre-processing (useful for cropping step). Cannot exceed 3.}

\item{textcol}{Character. Colour of text showing image date (can also be hexadecimal)}

\item{...}{Arguments to be passed to \code{\link{plotRGB}}}

\item{xr/xg/xb}{RasterBrick. Image time series bricks representing the red, green and blue channel, respectively}

\item{nc/nr}{Numeric. Number of columns and rows to plot, respectively. If the number of layers is greater than \code{nc*nr}, a screen prompt will lead to the next series of plots. These cannot exceed 4.}
}
\value{
\code{NULL} if \code{ggplot = FALSE} or an object of class \code{ggplot} if \code{ggplot = TRUE}, with the side effect of time series chips being plotted in both cases.
}
\description{
Plot image TS chips from 3 raster bricks (as RGB composites) based on a location, area buffer, and time buffer
}
\author{
Ben DeVries
}
\references{
Cohen, W. B., Yang, Z., Kennedy, R. (2010). Detecting trends in forest disturbance and recovery using yearly Landsat time series: 2. TimeSync - Tools for calibration and validation. Remote Sensing of Environment, 114(12), 2911-2924.
}

