% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tsComp.R
\name{tsComp}
\alias{tsComp}
\title{Image chip composites}
\usage{
tsComp(x, loc, fun = mean, na.rm = TRUE, start = NULL, end = NULL,
  buff = 17, cols = "PiYG", nbks = 35, nc = 3, nr = 3, plot = FALSE,
  plotlab = "data", exportChips = FALSE, exportZoo = FALSE, show = TRUE)
}
\arguments{
\item{x}{RasterBrick. Image time series brick}

\item{loc}{Location. Can be a vector of length 2 representing the x,y coordinates, or a SpatialPolygons or SpatialPoints object of \code{nrow = 1} (the first row will be taken if \code{nrow(loc) > 1}), or an extent object (which will be extended if a buffer > 0 is given; see below)}

\item{fun}{Function to apply over each pixel for each year. See \code{\link{annualSummary}} for more info.}

\item{na.rm}{Logical. Remove NA's when calculating fun over each year? If \code{fun} is a function that does not take a \code{na.rm} value, set \code{na.rm} to \code{NULL}.}

\item{start}{Numeric. OptionaL: earliest year to display.}

\item{end}{Numeric. Optional: latest year to display.}

\item{buff}{Numeric. Number of pixels to buffer the location in all directions. A higher buffer will essentially zoom out.}

\item{cols}{Character. Name of colour map to use (see display.brewer.all()) or a character vector with two or more colour names or hexadecimal values (as strings) between which to interpolate.}

\item{nbks}{Numeric. Number of breaks in the colour map}

\item{plot}{Logical. Plot pixel time series?}

\item{plotlab}{Character. y-axis label for the pixel time series plot.}

\item{exportChips}{Logical. Export processed chips to workspace as a rasterBrick?}

\item{exportZoo}{Logical. Export pixel time series as a zoo object?}

\item{show}{Logical. Plot the chips? Can be set to \code{FALSE} if you just want to export the chips as rasterBrick with or without the ggplot object.}

\item{nc/nr}{Numeric. Number of columns and rows to plot, respectively. If the number of layers is greater than \code{nc*nr}, a screen prompt will lead to the next series of plots. These cannot exceed 4.}
}
\description{
Display annual composites of image chips
}
\references{
Cohen, W. B., Yang, Z., Kennedy, R. (2010). Detecting trends in forest disturbance and recovery using yearly Landsat time series: 2. TimeSync - Tools for calibration and validation. Remote Sensing of Environment, 114(12), 2911-2924.
}

