% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tsCompRGB.R
\name{tsCompRGB}
\alias{tsCompRGB}
\title{Image chip composites (RGB version)}
\usage{
tsCompRGB(xr, xg, xb, loc, fun = mean, na.rm = TRUE, start = NULL,
  end = NULL, buff = 17, nc = 3, nr = 3, plot = FALSE,
  exportChips = FALSE, exportZoo = FALSE, textcol = "white",
  plotlabs = c("red", "green", "blue"), ...)
}
\arguments{
\item{loc}{Location. Can be a vector of length 2 representing the x,y coordinates, or a SpatialPolygons or SpatialPoints object of \code{nrow = 1} (the first row will be taken if \code{nrow(loc) > 1}), or an extent object (which will be extended if a buffer > 0 is given; see below)}

\item{fun}{Function to apply over each pixel for each year. See \code{\link{annualSummary}} for more info.}

\item{na.rm}{Logical. Remove NA's when calculating fun over each year? If \code{fun} is a function that does not take a \code{na.rm} value, set \code{na.rm} to \code{NULL}.}

\item{start}{Numeric. OptionaL: earliest year to display.}

\item{end}{Numeric. Optional: latest year to display.}

\item{buff}{Numeric. Number of pixels to buffer the location in all directions. A higher buffer will essentially zoom out.}

\item{plot}{Logical. Plot individual band time series?}

\item{exportChips}{Logical. Export processed chips to workspace as a list of rasterBricks (R, G, B)?}

\item{exportZoo}{Logical. Export pixel time series as \code{zoo} objects?}

\item{textcol}{Character. Colour of text showing image date (can also be hexadecimal)}

\item{plotlabs}{Character. Vector of length 3 indicating the labels for each of the zoo plots (if \code{plot=TRUE})}

\item{...}{Arguments to be passed to \code{\link{plotRGB}}}

\item{xr/xg/xb}{RasterBrick. Image time series bricks representing the red, green and blue channel, respectively}

\item{nc/nr}{Numeric. Number of columns and rows to plot, respectively. If the number of layers is greater than \code{nc*nr}, a screen prompt will lead to the next series of plots. These cannot exceed 4.}

\item{show}{Logical. Show image chips? Set to \code{FALSE} if you just want to export them to rasterBricks and/or export the \code{ggplot} object without viewing the chips.}
}
\value{
\code{NULL} if \code{exportZoo = FALSE} and \code{exportChips = FALSE} or a list of \code{zoo} time series objects if \code{exportZoo = TRUE}, or a list of subsetted and cropped rasterBricks if \code{exportChips = TRUE}, with the side effect of time series chips being plotted in all cases. If both \code{exportChips} and \code{exportZoo} are \code{TRUE}, a list consisting both lists will be returned.
}
\description{
Plot annual image composites chips from 3 raster bricks (as RGB composites) based on a location, area buffer, and time
}
\author{
Ben DeVries
}
\references{
Cohen, W. B., Yang, Z., Kennedy, R. (2010). Detecting trends in forest disturbance and recovery using yearly Landsat time series: 2. TimeSync - Tools for calibration and validation. Remote Sensing of Environment, 114(12), 2911-2924.
}

