% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tsChipsRGB.R
\name{tsChipsRGB}
\alias{tsChipsRGB}
\title{Plot TS Chips (RGB version)}
\usage{
tsChipsRGB(xr, xg, xb, loc, start = NULL, end = NULL, buff = 17,
  percNA = 20, nc = 3, nr = 3, plot = FALSE, exportChips = FALSE,
  exportZoo = FALSE, textcol = "white", plotlabs = c("red", "green",
  "blue"), ...)
}
\arguments{
\item{loc}{Location. Can be a vector of length 2 representing the x,y coordinates, or a SpatialPolygons or SpatialPoints object of \code{nrow = 1} (the first row will be taken if \code{nrow(loc) > 1}), or an extent object (which will be extended if a buffer > 0 is given; see below)}

\item{start}{Date. OptionaL: earliest date ("yyyy-dd-mm") to display.}

\item{end}{Date. Optional: latest date ("yyyy-dd-mm") to display.}

\item{buff}{Numeric. Number of pixels to buffer the location in all directions. A higher buffer will essentially zoom out.}

\item{percNA}{Numeric. Maximum allowable \% NA in the cropped image chips}

\item{plot}{Logical. Plot individual band time series?}

\item{exportChips}{Logical. Export processed chips to workspace as a list of rasterBricks (R, G, B)?}

\item{exportZoo}{Logical. Export pixel time series as \code{zoo} objects?}

\item{textcol}{Character. Colour of text showing image date (can also be hexadecimal)}

\item{plotlabs}{Character. Vector of length 3 indicating the labels for each of the zoo plots (if \code{plot=TRUE})}

\item{...}{Arguments to be passed to \code{\link{plotRGB}}}

\item{xr/xg/xb}{RasterBrick. Image time series bricks representing the red, green and blue channel, respectively}

\item{nc/nr}{Numeric. Number of columns and rows to plot, respectively. If the number of layers is greater than \code{nc*nr}, a screen prompt will lead to the next series of plots. These cannot exceed 4.}

\item{show}{Logical. Show image chips? Set to \code{FALSE} if you just want to export them to rasterBricks and/or export the \code{ggplot} object without viewing the chips.}
}
\value{
\code{NULL} if \code{ggplot = FALSE} or an object of class \code{ggplot} if \code{ggplot = TRUE}, with the side effect of time series chips being plotted in both cases. If \code{export = TRUE}, a list of objects of class rasterBrick, and if both \code{ggplot} and \code{export} are \code{TRUE}, a list including a list of rasterBricks and a ggplot object.
}
\description{
Plot image TS chips from 3 raster bricks (as RGB composites) based on a location, area buffer, and time buffer
}
\author{
Ben DeVries
}
\references{
Cohen, W. B., Yang, Z., Kennedy, R. (2010). Detecting trends in forest disturbance and recovery using yearly Landsat time series: 2. TimeSync - Tools for calibration and validation. Remote Sensing of Environment, 114(12), 2911-2924.
}

